follow_recall/requirements.txt
aiohttp==3.8.1
click==8.1.3
python-dateutil==2.8.2
pyssdb==0.4.2
tqdm==4.64.0
pymongo==4.1.1
redis==4.3.1
requests==2.27.1
psycopg2-binary==2.9.3
